+++ packages/browser/src/index.ts
@@ -77,4 +77,4 @@ export type { Span } from '@sentry/core';
 export { makeBrowserOfflineTransport } from './transports/offline';
 export { browserProfilingIntegration } from './profiling/integration';
 export { spotlightBrowserIntegration } from './integrations/spotlight';
-export { browserSessionIntegration } from './integrations/session';
+export { browserSessionIntegration } from './integrations/browsersession';
