+++ packages/browser/src/integrations/browsersession.ts
